YAN-1633 Split big SQL DDL into individual files
Each file does a bit of what the old file used to do in its entirety.
However, having this separation will allow us to reuse these files in
other situations (e.g., if we want to create the tables in a different
schema).

With this new split, we now also prepare the databse for testing in two
steps: we first create the ska_dlm user and the database, and then on a
second step, using that new user and database, we create the tables.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>