YAN-1633 Split big SQL DDL into individual files
Each file does a bit of what the old file used to do in its entirety,
which makes them more useful (e.g., if we want to only create tables and
not a user).  Additionally, there is no explicit reference to the public
schema anymore, which makes the relations definitions more reusable in
other setups.

With this new split, we now also prepare the databse for testing in two
steps: we first create the ska_dlm user and the database, and then on a
second step, using that new user and database, we create the tables.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>