YAN-1660 Introduce db_access module
This new module contains the necessary classes to centralise the logic
of building URLs and contacting PostgREST in order to perform certain
types of queries. On top of that, a globally available, pre-initialised
object called DB is made available for the rest of the modules to use as
a convenience. Queries are built internally in the PostgRESTAccess
class, and exceptions are wrapped in the DBQueryError exception class,
which is raised instead for uniformity.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>