YAN-1660 Remove REST.cfg_file from config.yaml
The location of the PostREST configuration file is of no concern to the
ska-dlm libraries, which are clients to PostREST (and in fact the
preference was never used, and its default value pointed to a file that
didn't exist). Let's therefore remove it in favour of a cleaner, simpler
configuration.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>