YAN-1660 Remove DB section from configuration file
All the ska-dlm libraries talk exclusively to PostgREST at this point in
time, so there is no need to configuration direct database access for
the ska-dlm libraries (and in fact this preference was indeed never
used).

This commit removes the DB section from the configuration file used by
the ska-dlm libraries, and any further references to a separate secrets
file. When/if we update these libraries to talk to the DB directly we
can reinstate these preferences as necessary.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>