YAN-1660 Remove "report" parameters for consistent error reporting
The "report" parameter found in some of the functions was a source of
inconsistent reporting, since logging would only sometimes be issued.
Moreover, it showed that some programming issues were not addressed,
like the fact that certain parameter combinations was required by some
functions, which were happily working when not correctly invoked.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>