YAN-1660 Use new DB object to perform DB queries
Using this new object means we no longer perform URL building and
concatenation ourselves (which was parallel to building an SQL query
manually as a string). Instead all queries are now defined in terms of
HTTP parameters and POST data, as appropriate.

I also took the chance to clean up parameter validation in general, and
handling of internal errors around DB operations. In both situations
errors were simply logged and a sentinel empty value was being returned,
instead of an exception being raised. I've changed most places I've
found (but there could be more) to raise an exception as early as
possible. Due to this, some tests needed adjusting, so they have been
changed to expect exceptions, and in some places also augmented to check
that certain updates in the database are reflected in subsequent
queries.