YAN-1662 De-duplicate CLI entry point definition
Now that we are exposing more functions as CLI entry points, the code is
starting to explode. Having a function that implements all the common
steps is thus desirable.

Additionally, the new function prints a nice "caused by"-chained
exception trail for those exceptions that we indicate as needing a
simple error message.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>