YAN-1633: Deploy helm chart in CI
Add make targets for setting up and deploying the DLM. Update to the
latest .make reference to give access to the latest Makefile templates
despite only using it sparingly (we want to minimise our dependencies on
the observed brittle nature of it) primarily so we can easily debug
deployments in CI using `make k8s-namespace-credentials`.

We then use these new targets to deploy, but not use, the chart in CI.
In a future commit we will alter out tests to target this instance
instead of the processes manually started using `setup_services.sh`.