YAN-1633: Run tests against helm chart
Instead of having the test infrastructure spread across multiple
locations (`.gitlab-ci.yml`, `scripts/`,
`test-services.docker-compose.yml`) we consolidate them into the Helm
chart and run the tests against this instance. This requires a bit of
trickery to ensure we hit the correct endpoints as currently the config
is hardcoded into the source code and we have to monkey-patch it in the
test code in order to hit the correct endpoint at test time. Making this
more dynamic should be the focus of a future ticket.

Additionally this change removes the DLM storage manager from starting
with the other services. After discussions with the team, it was decided
that two Helm charts would be created, one for pure infrastructure (the
current chart), and another to hold all the DLM daemons and services,
which would _consume_ the infrastructure chart. To keep this change
small, this is work that will be performed in the future.