YAN-2011 Read postrest configuration from Vault
This support is implemented by creating a configuration file that we can
point the postgrest binary to. The Vault injection mechanism can only
create files, and PostgREST favours environment variables over
configuration files. Therefore we need to assume that certain
configuration keys will be present in the Vault, so we can exclude the
corresponding environment variables from the container definition.

An alternative solution would be to generate a shell script that we can
source, and that exports environment variables. However, the postgrest
image we use doesn't come with a shell.

This support is optional and turned off by default.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>