YAB-41 Configure PostgREST via Secret and VSO
The DLM chart previously used Vault Agent Inject to read secrets from
Vault and create a file on disk with the URI to the PostegreSQL server,
which the PostgREST executable could read. Usage of the Vault Agent
Injector (VAI) is deprecated though, with everyone encouraged to move to
using the Vault Secrets Operator (VSO).

This commit removes support for VAI, replacing it with VSO. VSO is
leveraged by creating a VaultStaticSecret object, which in turn creates
a Secret that the PostgREST pod can consume. In order to make the Secret
more flexible, its contents are now more fine grained, with the host,
username, password and database being listed as separate keys. This
requires the chart to build the PostegreSQL URI from those separate
bits, but on the other side it allows more flexibility in the long term.
These Secret key names are modelled after the environment variables that
libpq accepts. When using VSO, the Vault path, mount point and engine
type have to be given as parameters.

In order to accommodate for the deployment strategy that AIV is pursuing
in AA Low, a Helm value can alternatively be directed to skip creating
this secret, and instead use a pre-existing one, which is pointed by
name. The provided Secret needs to provide the same keys as the one
generated by the DLM chart, deployment will fail if this requirement is
not fulfilled.

In development environments the chart is installed and configured to
spawn its own PostegreSQL server. In these cases the Secret is created
to automatically point to this internal server.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>