YAB-41 Disable Vault in the dev and integration environments
In both cases we use an embedded PG database, and thus we shouldn't
attempt to obtain the secret from the Vault, and instead we should let
the chart generate it automatically. As discussed in !103 the better
solution is to rework the chart so we don't have to disable the Vault in
these cases.

Additionally the staging Helm values file points to an external
database. We don't use this file because we know the staging deployment
to be broken for other reasons, but we should at least strive to keep it
relevant for when the staging installation can be performed again.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>