YAB-41 Simplify secret setup for dev envirnoments
The initial implementation of the logic to create or not the DB
authentication Secret for PostgREST didn't cover well for the normal
developer experience, as users had to disable Vault explicitly when
using the internal PostgreSQL server.

This commit gives priority to _always_ auto-generate the Secret when the
internal PostgreSQL server is enabled. This in turn means that when
checking whether pre-existing Secrets, or Secrets derived from Vault
work, we cannot do it with an internal PostgreSQL server running.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>